+++ providers/prompt/darkPrompt.css
